+++ src/dynamic/check.rs
@@ -1,3 +1,5 @@
+use std::collections::HashSet;
+
 use indexmap::IndexMap;
 
 use crate::dynamic::{
@@ -224,17 +226,18 @@ impl SchemaInner {
                 // through referenced Input Objects, at least one of the
                 // fields in the chain of references must be either a
                 // nullable or a List type.
-                self.check_input_object_reference(&obj.name, &obj)?;
+                self.check_input_object_reference(&obj.name, &obj, &mut HashSet::new())?;
             }
         }
 
         Ok(())
     }
 
-    fn check_input_object_reference(
-        &self,
+    fn check_input_object_reference<'a>(
+        &'a self,
         current: &str,
-        obj: &InputObject,
+        obj: &'a InputObject,
+        ref_chain: &mut HashSet<&'a str>,
     ) -> Result<(), SchemaError> {
         fn typeref_nonnullable_name(ty: &TypeRef) -> Option<&str> {
             match &ty.0 {
@@ -255,7 +258,13 @@ impl SchemaInner {
                     .get(field.ty.type_name())
                     .and_then(Type::as_input_object)
                 {
-                    self.check_input_object_reference(current, obj)?;
+                    // don't visit the reference if we've already visited it in this call chain
+                    //  (prevents getting stuck in local cycles and overflowing stack)
+                    //  true return from insert indicates the value was not previously there
+                    if ref_chain.insert(this_name) {
+                        self.check_input_object_reference(current, obj, ref_chain)?;
+                        ref_chain.remove(this_name);
+                    }
                 }
             }
         }
@@ -475,6 +484,21 @@ mod tests {
             .field(InputValue::new("bottom", TypeRef::named_nn("BotLevel")));
         let bot_level = InputObject::new("BotLevel")
             .field(InputValue::new("top", TypeRef::named_nn("TopLevel")));
+        let schema = base_schema()
+            .register(top_level)
+            .register(mid_level)
+            .register(bot_level);
+        schema.finish().unwrap_err();
+    }
+
+    #[test]
+    fn test_recursive_input_objects_local_cycle() {
+        let top_level = InputObject::new("TopLevel")
+            .field(InputValue::new("mid", TypeRef::named_nn("MidLevel")));
+        let mid_level = InputObject::new("MidLevel")
+            .field(InputValue::new("bottom", TypeRef::named_nn("BotLevel")));
+        let bot_level = InputObject::new("BotLevel")
+            .field(InputValue::new("mid", TypeRef::named_nn("MidLevel")));
         let schema = base_schema()
             .register(top_level)
             .register(mid_level)