+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [5.0.9] 2023-05-25
+
+- Prevent input check stack overflow [#1293](https://github.com/async-graphql/async-graphql/pull/1293)
+- Change batch requests to run concurrently [#1290](https://github.com/async-graphql/async-graphql/issues/1290)
+
 # [5.0.8] 2023-05-09
 
 - Improve documentation about DataloaderImprove documentation about Dataloader [#1282](https://github.com/async-graphql/async-graphql/pull/1282)